Reset Update.kt change.
